requirements.txt
# EVA API Platform - Production Dependencies

# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Authentication & Authorization
python-jose[cryptography]==3.3.0
azure-identity==1.15.0
PyJWT==2.8.0

# Azure Services
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
openai==1.12.0  # Azure OpenAI SDK

# Database & Cache
redis==5.0.1

# HTTP & Utilities
httpx==0.26.0
python-multipart==0.0.9
slowapi==0.1.9  # Rate limiting

# GraphQL (Phase 3)
strawberry-graphql==0.219.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Rate Limiting
slowapi==0.1.9

# Validation & Serialization
email-validator==2.1.0
python-dateutil==2.8.2
